Divide code into different components
